Reimplemented basic spawn syscall.

+++ src/libs/std/functions/stdlib/exit.c
@@ -4,7 +4,7 @@
 
 _NORETURN void exit(int status)
 {
-    SYSCALL(SYS_PROCESS_EXIT, 1, status);
+    SYSCALL(SYS_PROCESS_EXIT, 1, (uint64_t)status);
     while (1)
     {
         asm volatile("ud2");
